move a bunch of stuff to components, set up a test store
TODO: use first feature as a demo
