+++ Readme.md
@@ -1,4 +1,5 @@
 The front end for timewinder
+run: npm start
 
 TODO:
 redux setup
