+++ src/components/postgame/post.js
