--- src/home/home.css
@@ -1,72 +1,0 @@
-.Home h1 {
-  color: #006666;
-  font-weight: 700;
-  font-size: 70px;
-  text-align: center;
-  margin: 0px;
-}
-.Home h3 {
-  margin: 0px;
-  text-align: center;
-}
-.Home .content {
-  width: 600px;
-  margin-left: auto;
-  margin-right: auto;
-  top: 200px;
-  position: relative;
-}
-.Home .footer {
-  position: absolute;
-}
-.Home #quote {
-  width: 100%;
-  height: 50px;
-  margin: 15px 0;
-  text-align: center;
-  font-size: 14px;
-}
-.Home #error {
-  margin-top: 5px;
-  text-align: center;
-  color: #FF0000;
-}
-.Home #lookup {
-  height: 45px;
-}
-.Home #searchServer {
-  margin: auto;
-  float: left;
-  font-size: 14px;
-  width: 70px;
-  height: 100%;
-}
-.Home #searchField {
-  margin: auto;
-  float: left;
-}
-.Home #searchButton {
-  margin: auto;
-  float: right;
-  width: 94px;
-  height: 100%;
-}
-.Home #searchLoader {
-  float: none;
-  font-size: 30px;
-  margin: 72px auto;
-}
-.Home .textfield {
-  width: 435px;
-  height: 100%;
-  font-size: 12pt;
-}
-.Home #champion {
-  width: 300px;
-}
-.Home #champion .blue {
-  float: left;
-}
-.Home #champion .red {
-  float: right;
-}
