+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import Routes from './routes';
 
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import './common/timewinder.css';
+import './components/common/timewinder.css';
 
 ReactDOM.render(
   <Routes />,
